+++ packages/frontend/app/components/Badge.tsx
@@ -1,18 +1,19 @@
+import { cx } from 'class-variance-authority'
+import type { PaymentPointerStatus } from '~/generated/graphql'
+
 type BadgeProps = {
-  status: 'ACTIVE' | 'INACTIVE' | undefined | null
+  status: PaymentPointerStatus
 }
 
 export const Badge = ({ status }: BadgeProps) => {
-  const spanClassNames =
-    status === 'ACTIVE'
-      ? 'bg-green-200 text-green-800'
-      : 'bg-red-200 text-red-800'
-
-  if (!status) return 'No status'
-
   return (
     <span
-      className={`${spanClassNames} inline-flex items-center gap-x-1.5 rounded-full  px-2 py-1 text-xs font-medium `}
+      className={cx(
+        'inline-flex items-center gap-x-1.5 rounded-full  px-2 py-1 text-xs font-medium',
+        status === 'ACTIVE'
+          ? 'bg-green-200 text-green-800'
+          : 'bg-red-200 text-red-800'
+      )}
     >
       {status}
     </span>
